ДЗ4. Углубление в БД

See merge request KaznacheevPavel/internship-restaurant!5

+++ common-lib/src/main/java/ru/kaznacheev/restaurant/common/exception/OrderNotFoundException.java
@@ -1,17 +1,19 @@
 package ru.kaznacheev.restaurant.common.exception;
 
-import ru.kaznacheev.restaurant.common.dto.response.ResponseTitle;
+import ru.kaznacheev.restaurant.common.dto.exception.ResponseTitle;
 
 /**
  * Исключение, выбрасываемое при попытке получения несуществующего заказа.
  */
 public class OrderNotFoundException extends BaseException {
+
     /**
-     * Создает экземпляр класса {@link OrderNotFoundException} с указанными значениями для {@link BaseException}.
+     * Конструктор исключения.
      *
      * @param id Идентификатор заказа, который вызвал исключение
      */
-    public OrderNotFoundException(int id) {
+    public OrderNotFoundException(Long id) {
         super(ResponseTitle.NOT_FOUND.name(), ResponseTitle.NOT_FOUND.getStatus(), "Заказ с идентификатором " + id + " не найден");
     }
+
 }
